refactor(types): User 타입을 shared layer로 이동
- User 타입을 entities에서 shared로 이동하여 중복 제거
- TypeScript 타입 에러 해결
