+++ src/entities/user/types/index.ts
@@ -1,0 +1,1 @@
+export type { User } from '../../../shared/types/common';
